+++ demisto_sdk/commands/common/docker_helper.py
@@ -336,7 +336,11 @@ class DockerBase:
                 )
             except (docker.errors.BuildError, docker.errors.APIError, Exception) as e:
                 errors = str(e)
-                logger.critical(f"{log_prompt} - Build errors occurred: {errors}")
+                error_log = f"{log_prompt} - Build errors occurred: {errors}"
+                if errors == "'javascript'":
+                    logger.warning(error_log)
+                else:
+                    logger.critical(error_log)
 
         return test_docker_image, errors