+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Changelog
 ## Unreleased
+* updated the **run-unit-tests** command to not fail on JavaScript items, but skip them instead.
 * Added the *auto-replace-uuids* flag to the **download** command. set this flag to False to avoid UUID replacements when downloading using download command.
 * **format** command will run without the content graph if graph creation fails.
 * Internal: Replaced the `tools._read_file` function with a more generic `tools.safe_read_unicode` function.